# openapi.yaml
openapi: 3.1.0
info:
  title: RightFlow API
  description: |
    RightFlow is a Hebrew PDF form management platform with AI-powered field extraction,
    workflow automation, and multi-channel delivery.

    ## Authentication
    Most endpoints require a **Clerk JWT** token passed as a Bearer token:
    ```
    Authorization: Bearer <your_clerk_token>
    ```

    ## Roles
    - **worker** – Submit and view forms
    - **manager** – Manage webhooks, analytics, connectors
    - **admin** – Full access including user management

    ## Multi-tenancy
    All authenticated requests are scoped to the user's organization automatically.
  version: 3.0.0
  contact:
    name: RightFlow Support
    url: https://rightflow.railway.app
servers:
  - url: https://rightflow.railway.app
    description: Production
  - url: http://localhost:3003
    description: Local development
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk session token obtained from the frontend SDK.
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      example:
        error: FORBIDDEN
        message: Insufficient permissions
...
    FormField:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, checkbox, radio, dropdown, signature, email, tel, date, number, textarea, select]
        required:
          type: boolean
        pageNumber:
          type: integer
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        options:
          type: array
          items:
            type: string
    Form:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
        settings:
          type: object
        status:
          type: string
          enum: [draft, active, archived]
        isActive:
          type: boolean
        organizationId:
          type: string
          format: uuid
        createdById:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Submissions ───────────────────────────────────
...
paths:
...
  /api/v1/forms:
    get:
      tags: [Forms]
      summary: List forms
      security:
        - ClerkAuth: []
      parameters:
        - name: isActive
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        '200':
          description: List of forms
          content:
            application/json:
              schema:
                type: object
                properties:
                  forms:
                    type: array
                    items:
                      $ref: '#/components/schemas/Form'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags: [Forms]
      summary: Create form
      security:
        - ClerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                description:
                  type: string
                fields:
                  type: array
                  items:
                    $ref: '#/components/schemas/FormField'
                settings:
                  type: object
                status:
                  type: string
                  enum: [draft, active, archived]
      responses:
        '201':
          description: Created form
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Form'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/webhooks/grow:
    post:
      tags: [Billing]
      summary: Grow payment webhook (incoming)
      description: |
        Receives payment notifications from the Grow payment provider.
        - Signature verified via HMAC
        - Rate limited: 100 req/min per IP
        - Idempotent: duplicate transaction IDs are ignored
        - Optional IP whitelist via `GROW_IP_WHITELIST` env var
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Grow webhook payload (provider-defined schema)
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
...